+++ uniflow-test/src/test/kotlin/io/uniflow/test/impl/AbstractSampleFlow.kt
@@ -22,7 +22,7 @@ abstract class AbstractSampleFlow(val defaultState: UIState) : DataFlow, DataPub
     internal val defaultDataPublisher = simpleListPublisher(defaultState, "$tag-Publisher")
     override suspend fun publishState(state: UIState, pushStateUpdate: Boolean) = defaultPublisher().publishState(state, pushStateUpdate)
     override suspend fun publishEvent(event: UIEvent) = defaultPublisher().publishEvent(event)
-    override suspend fun getState(): UIState = defaultPublisher().getState()
+    override fun getState(): UIState = defaultPublisher().getState()
     override fun defaultPublisher(): DataPublisher = defaultDataPublisher
 
     private val actionReducer = ActionReducer(::defaultPublisher, coroutineScope, UniFlowDispatcher.dispatcher.io(), Channel.BUFFERED, tag)